refined some of the classes

+++ InventoryManagement/src/main/java/org/zurika/inventorymanagement/controller/OrderController.java
@@ -4,6 +4,7 @@ import java.util.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.zurika.inventorymanagement.exception.ResourceNotFoundException;
 import org.zurika.inventorymanagement.model.Order;
 import org.zurika.inventorymanagement.service.OrderService;
 
@@ -21,7 +22,10 @@ public class OrderController {
 
     @GetMapping("/{id}")
     public Optional<Order> findOrderById(@PathVariable Long id){
-        return orderService.findById(id);
+        return Optional.of(orderService.findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException
+        ("Order not found with id: " + id)
+        ));
     }
 
     public Order createOrder(@RequestBody Order order){
@@ -30,6 +34,10 @@ public class OrderController {
 
     @DeleteMapping("/{id}")
     public void deleteOrder(@PathVariable Long id){
+        if(!orderService.findById(id).isPresent()){
+            throw new ResourceNotFoundException(
+                "Order not found with id: " + id);
+        }
         orderService.deleteById(id);
     }
 }
