+++ InventoryManagement/src/main/java/org/zurika/inventorymanagement/controller/ProductController.java
@@ -2,7 +2,9 @@ package org.zurika.inventorymanagement.controller;
 
 import java.util.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.*;
 import org.springframework.web.bind.annotation.*;
+import org.zurika.inventorymanagement.exception.ResourceNotFoundException;
 import org.zurika.inventorymanagement.model.Product;
 import org.zurika.inventorymanagement.service.ProductService;
 
@@ -14,15 +16,28 @@ public class ProductController {
     private ProductService productService;
 
     @GetMapping
-    public List<Product> getAllProducts() {
-        return productService.findAll();
+    public Page<Product> getAllProducts(Pageable pageable) {
+        return productService.findAll(pageable);
     }
 
     @GetMapping("/{id}")
     public Optional<Product> getProductByid(@PathVariable Long id) {
-        return productService.findById(id);
+        return Optional.of(productService.findById(id).orElseThrow(
+            () -> new ResourceNotFoundException(
+                "Product not found with id: " + id)
+        ));
     }
 
+    // searching and filtering products using pagination
+    @GetMapping("/search")
+    public Page<Product> searchProduct(
+        @RequestParam(required = false) String category,
+        @RequestParam(required = false) String name,
+        Pageable pageable
+    ){
+        return productService.search(category, name, pageable);
+    }
+    
     @PostMapping
     public Product createProduct(@RequestBody Product product){
         return productService.save(product);
@@ -30,7 +45,10 @@ public class ProductController {
 
     @DeleteMapping("/{id}")
     public void deleteProduct(@PathVariable Long id){
+        if(!productService.findById(id).isPresent()){
+            throw new ResourceNotFoundException(
+                "Product not found with id: " + id);
+        }
         productService.deleteById(id);
-    }
-    
+    }    
 }
